bugId:36218; changes after code review; WIP; TODO: updating visualization table/cache
git-svn-id: svn+ssh://svn.wikia-code.com/svn/wikia/trunk@56119 e846bd91-c042-42a7-a218-4b65e2c1904d

+++ maintenance/wikia/ImageReview/AdminUpload/remove.php
@@ -3,7 +3,9 @@
  * How to run this script:
  * $ cd maintenance/wikia/ImageReview/AdminUpload/upload.php
  * To remove an image from wikia.com as a WikiaBot:
- * $ SERVER_ID=80433 php remove.php --conf /usr/wikia/docroot/wiki.factory/LocalSettings.php --imagename="Wikia-Visualization-Add-3,starwars.jpg" --userid=4663069
+ * $ SERVER_ID=80433 php remove.php --conf /usr/wikia/docroot/wiki.factory/LocalSettings.php --imagename="Wikia-Visualization-Add-3,starwars.jpg" --userid=4663069 --wikiid=147
+ * Optional parameter
+ * --overwritelang=en
  */
 require_once("../../../commandLine.inc");
 
@@ -13,32 +15,40 @@ $user = F::build('User', array($userId), 'newFromId');
 
 if( !($user instanceof User) ) {
 	echo 'ERROR: Could not get user object'."\n";
-	exit;
+	exit(1);
 }
 
 if( !$user->isAllowed('delete') ) {
 	echo 'ERROR: You do not have right permissions'."\n";
-	exit;
+	exit(1);
 }
 
 if( empty($imageName) ) {
 	echo 'ERROR: Invalid image name'."\n";
-	exit;
+	exit(1);
 }
 
 $imageTitle = Title::newFromText($imageName, NS_FILE);
 
 if( !($imageTitle instanceof Title) ) {
 	echo 'ERROR: Could not get title object';
-	exit;
+	exit(1);
 }
 
 $file = wfFindFile($imageTitle);
-$status = $file->delete('automated deletion');
+
+if( $file instanceof File && $file->exists() ) {
+	$status = $file->delete('automated deletion');
+} else {
+	$status->ok = false;
+	$status->errors = array('ERROR: File does not exist');
+}
 
 if( $status->ok === true ) {
 	echo "\n"."INFO: File has been deleted"."\n";
+	exit(0);
 } else {
 	echo "\n"."ERROR: File has not been deleted"."\n";
 	var_dump($status);
+	exit(1);
 }
